Tiberius no longer defaults to Derpibooru
Tiberius now requires a target domain be specified in the constructor.

Filter Aliases can now be used when setting Filters for any domain.

Filter aliases are case insensitive names and resolved by asking the
target domain for it's system filters.

Tiberius defaults to trying to use the filter alias 'default'